+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3extralearners 0.6.0-9000
 
+* BREAKING CHANGE: lightgbm's early stopping mechanism now uses the task's test set.
+
 # mlr3extralearners 0.6.0
 
 * Feat: Added learner `LearnerClassifGlmer` (https://github.com/mlr-org/mlr3extralearners/issues/243)